Implemented Accelerations
Accelerations are the association of a Key and ModMask combination with
a specific closure function. This system provides a means of tracking
and configuring per-application-window accelerators in a dynamic way.
The StockItems within CTK provide the standard key combinations such as
the "_Quit" StockItem has a <Ctrl>q key combination assigned. Developers
can rely upon those defaults or override them in a number of ways.

* AccelMap: display-singleton for managing accelerator mappings
* Accelerator: specific AccelMap entries
* AccelKey: tracking specific AccelGroupEntry key combinations
* AccelGroupEntry: specific AccelGroup item
* AccelGroup: grouping of accelerator handlers (closures)
* AccelGroups: managing multiple AccelGroups